Applying tropikhajma's patch to fix KStars' build issue due to log10 on Sun Studio Express / Solaris.
Continues to build fine on i386 after patching.

Thanks for the patch!

CCMAIL: kstars-devel@kde.org
BUG: 200093


svn path=/trunk/KDE/kdeedu/kstars/; revision=996428

+++ kstars/skycomponents/starcomponent.cpp
@@ -263,7 +263,7 @@ float StarComponent::zoomMagnitudeLimit() const {
     // Reducing the slope w.r.t zoom factor to avoid the extremely fast increase in star density with zoom
     // that 4.444 gives us (although that is what the derivation gives us)
 
-    float maglim = 3.7 * ( lgz - lgmin ) + 2.222 * log10( Options::starDensity() ) + 3.5;
+    float maglim = 3.7 * ( lgz - lgmin ) + 2.222 * log10( static_cast<float>(Options::starDensity()) ) + 3.5;
 
     return maglim;
